requirements.txt
attrs==19.3.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
bz2file==0.98
certifi==2019.9.11
cloudpickle==1.2.2
configparser==4.0.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
h5py==2.10.0
html5lib==0.9999999
imageio==2.6.1
ipaddress==1.0.23
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.10.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
mock==3.0.5
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.2
nibabel==2.5.1
notebook==5.7.8
numpy==1.16.5
pandas==0.24.2
pandocfilters==1.4.2
pathlib==1.0.1
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.10.0
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.5
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
pyzmq==18.1.0
qtconsole==4.5.5
scandir==1.10.0
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.2.2
Send2Trash==1.5.0
simplegeneric==0.8.1
SimpleITK==1.2.4
singledispatch==3.4.0.3
sip==4.19.13
six==1.13.0
sklearn==0.0
subprocess32==3.5.4
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
terminado==0.8.2
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1

--extra-index-url=https://packagecloud.io/github/git-lfs/pypi/simple